pyproject.toml
[tool.poetry]
name = "codex-bot"
version = "0.1.0"
description = ""
authors = ["CodeX"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
websockets = "^12.0"
aiohttp = "^3.9.3"
prometheus-client = "^0.20.0"
pytoniq-core = "^0.1.31"
pytonconnect = "^0.3.0"
qrcode = "^7.4.2"
redis = "^5.0.2"
aiogram = "^3.4.1"

[tool.poetry.scripts]
start-bot = "codex_bot.bot:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
